+++ tests/unit/dsp/platform.cpp
@@ -7084,7 +7084,6 @@ TEST(decodeRedfishActionPdr, GoodTest)
 }
 #endif
 
-#if HAVE_LIBPLDM_API_TESTING
 TEST(StateEffecterPDR, testExtractPossibleStates)
 {
     constexpr size_t pdrSize = sizeof(pldm_state_effecter_pdr) -
@@ -7162,9 +7161,7 @@ TEST(StateEffecterPDR, testExtractPossibleStates)
     EXPECT_EQ(bitLists[1][0], 0xAA);
     EXPECT_EQ(bitLists[1][1], 0x55);
 }
-#endif
 
-#if HAVE_LIBPLDM_API_TESTING
 TEST(StateEffecterPDR, testInvalidBufferTooSmall)
 {
     alignas(pldm_state_effecter_pdr) unsigned char
@@ -7182,9 +7179,7 @@ TEST(StateEffecterPDR, testInvalidBufferTooSmall)
 
     EXPECT_EQ(rc, -EOVERFLOW) << "Expected EOVERFLOW for small buffer";
 }
-#endif
 
-#if HAVE_LIBPLDM_API_TESTING
 TEST(StateEffecterPDR, testNullPointer)
 {
     int rc;
@@ -7198,9 +7193,7 @@ TEST(StateEffecterPDR, testNullPointer)
 
     EXPECT_EQ(rc, -EINVAL) << "Expected EINVAL for null pointer";
 }
-#endif
 
-#if HAVE_LIBPLDM_API_TESTING
 TEST(StateEffecterPDR, testZeroCount)
 {
     constexpr size_t pdrSize = sizeof(pldm_state_effecter_pdr);
@@ -7221,9 +7214,7 @@ TEST(StateEffecterPDR, testZeroCount)
     EXPECT_EQ(count, 0) << "Should not iterate when count is 0";
     EXPECT_EQ(rc, 0) << "Expected success (rc=0), got rc=" << rc;
 }
-#endif
 
-#if HAVE_LIBPLDM_API_TESTING
 TEST(StateEffecterPDR, testSingleEntry)
 {
     constexpr size_t pdrSize = sizeof(pldm_state_effecter_pdr) -
@@ -7282,9 +7273,7 @@ TEST(StateEffecterPDR, testSingleEntry)
     EXPECT_EQ(rc, 0);
     EXPECT_EQ(entryCount, 1);
 }
-#endif
 
-#if HAVE_LIBPLDM_API_TESTING
 TEST(StateEffecterPDR, testTruncatedEntry)
 {
     /* Entry claims size=5 but buffer only has room for 2 state bytes */
@@ -7319,9 +7308,7 @@ TEST(StateEffecterPDR, testTruncatedEntry)
 
     EXPECT_EQ(rc, -EOVERFLOW) << "Expected EOVERFLOW for truncated entry";
 }
-#endif
 
-#if HAVE_LIBPLDM_API_TESTING
 TEST(StateEffecterPDR, testZeroPossibleStatesSize)
 {
     constexpr size_t pdrSize =
@@ -7373,4 +7360,3 @@ TEST(StateEffecterPDR, testZeroPossibleStatesSize)
     EXPECT_EQ(rc, 0);
     EXPECT_EQ(entryCount, 1);
 }
-#endif